+++ RELEASE-NOTES.md
@@ -4,6 +4,11 @@ This is a pre-release version and APIs will change quickly. Before `1.0` release
 
 Please note after `1.0` Semver will be followed using normal protocols.
 
+# Version 0.10.2
+
+## Improvements
+* Added new keyword `bind` that will let you directly bind events without event delegation. This can be useful for attaching to events that do not bubble.
+
 # Version 0.10.1
 
 ## Improvements